* switched Unsafe.Hash algorithm from FNV-1a to Murmur2 * added support for function pointers in 2021.2+ to function wrappers (CastableAction, CastableEvent, etc) * added support for function pointers on AOT platforms for MethodInvocationHelper with specialized signatures * hotfix for non-development StringHashing.DumpReverseLookupStats * added CameraHelper.GetStateHash and TransformHelper.GetStateHash for quickly determining if their respective states have changed relative to a previous moment * fixed Unsafe.PinString(StringBuilder) to properly account for chained builders * various performance and garbage generation optimizations * added Reserve extension methods to sets, dictionaries, lists, and StringBuilders for ensuring capacity for new elements * added polyfills for various dictionary and set operations outside of net standard * added tick rate conversion functions to Profiling * added RegularPolyGraphic for rendering regular polygons with an arbitrary number of sides

+++ Assets/BeauUtil/Collections/OffsetLength.cs
@@ -360,7 +360,7 @@ namespace BeauUtil
 
         public override string ToString()
         {
-            return string.Format("[{0}({1})]", ((int) Offset).ToStringLookup(), ((int) Length).ToStringLookup());
+            return string.Format("[{0}({1})]", Offset.ToStringLookup(), Length.ToStringLookup());
         }
 
         #endregion // Overrides
@@ -478,7 +478,7 @@ namespace BeauUtil
 
         public override string ToString()
         {
-            return string.Format("[{0}({1})]", ((int) Offset).ToStringLookup(), ((int) Length).ToStringLookup());
+            return string.Format("[{0}({1})]", Offset.ToStringLookup(), Length.ToStringLookup());
         }
 
         #endregion // Overrides
